+++ scripts/lichessTools-init.js
@@ -123,6 +123,7 @@ lichessTools.loadTool(LiChessTools.Tools.TournamentTopBoard);
 lichessTools.loadTool(LiChessTools.Tools.BotFilters);
 lichessTools.loadTool(LiChessTools.Tools.AnalysisReadComments);
 lichessTools.loadTool(LiChessTools.Tools.CommentEditing);
+lichessTools.loadTool(LiChessTools.Tools.PlayerExtraButtons);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);